Update dependencies. Replace tslint % @1password/front-end-style
The 1Password frontend linters included a CSS linter that is not needed in this library. This change fully removes TSLint and implements the same ESlint plugins used by @1Password/front-end-style.

We also get a minor vuln fix for free: certain dependencies pulled in by @1Password/front-end-style, but not used by this library, required updates.